requirements.txt
albumentations==1.3.1
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
executing==1.2.0
filelock==3.12.2
fonttools==4.42.1
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.2
opencv-python-headless==4.8.0.76
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycocotools==2.0.7
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
qudida==0.0.4
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.8.25
timm==0.9.6
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers@07998ef39926b76d3f6667025535d0859eed61c3
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.2
wcwidth==0.2.6
plotly
scipy 
tensorboard 
umap-learn
